requirements.txt
tweepy
pillow 
datetime
